Started the QT gui.For now just button is shown on the screen

+++ Eter/Eter/CardType.h
@@ -1,4 +1,4 @@
 #pragma once
-#include <cstdint>
+#include "Includes.h"
 
-enum class CardType : uint8_t { MinionCard,ExplosionCard,HoleCard,MageCard,ElementalCard};
+enum class CardType : uint16_t { MinionCard,ExplosionCard,HoleCard,MageCard,ElementalCard};
