+++ Eter/Eter/MageType.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <cstdint>
 
-enum class MageType: uint8_t { FireMage,EarthMage,AirMage,WaterMage };
+enum class MageType: uint16_t { FireMage,EarthMage,AirMage,WaterMage };
